# outputs/deps-2026-09/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Self-hosted stock analysis platform for Thai and US markets",
    "title": "StockViz API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/stocks/names": {
      "get": {
        "description": "Batch name + market type lookup from the local DB.\n\nReturns ``{symbol: {name, market}}`` for every requested symbol.\nSymbols not found in the DB fall back to the raw symbol string with market=null.\nUsed by the Sidebar to display company names instead of tickers,\nand to distinguish FUND symbols (show NAV) from stocks (show price).",
        "operationId": "get_stock_names_api_v1_stocks_names_get",
        "parameters": [
          {
            "description": "Comma-separated list of symbols, e.g. PTT.BK,AAPL,NVDA",
            "in": "query",
            "name": "symbols",
            "required": true,
            "schema": {
              "description": "Comma-separated list of symbols, e.g. PTT.BK,AAPL,NVDA",
              "title": "Symbols",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Stock Names",
        "tags": [
          "stocks"
        ]
      }
    },
...
    "/api/v1/stocks/search": {
      "get": {
        "description": "Search stocks by symbol, English name, or Thai name.\n\nPriority order:\n  1. Exact symbol match  (highest relevance)\n  2. Symbol prefix match\n  3. Thai name (name_th) contains q\n  4. English name contains q\n  5. Yahoo Finance symbol search (fallback for unlisted symbols)",
        "operationId": "search_stocks_api_v1_stocks_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 50,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Stocks",
        "tags": [
          "stocks"
        ]
      }
    },
...
  }
}